Use generated data for tests instead of hardcoded
